sagemaker/requirements.txt
boto3==1.14.4
botocore==1.17.4
coloredlogs==14.0
cycler==0.10.0
docutils==0.15.2
humanfriendly==8.2
importlib-metadata==1.6.1
jmespath==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.18.5
packaging==20.4
Pillow==7.1.2
pkg-resources==0.0.0
protobuf==3.12.2
protobuf3-to-dict==0.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
s3transfer==0.3.3
sagemaker==1.64.1
scipy==1.4.1
six==1.15.0
smdebug-rulesconfig==0.1.4
urllib3==1.25.9
zipp==3.1.0
